requirements.txt
streamlit
pdf2image
pytesseract
reportlab
requests
python-dotenv
